panier

+++ src/Controller/ProductController.php
@@ -7,15 +7,18 @@ use App\Entity\Category;
 use App\Form\ProductType;
 use App\Form\CategoryType;
 use App\Controller\StoreController;
+use App\Controller\ProductController;
 use App\Repository\ProductRepository;
 use App\Repository\CategoryRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
-class StoreController extends AbstractController
+class ProductController extends AbstractController
 {
     // **********************************************
     // *********************PRODUIT******************
@@ -25,13 +28,10 @@ class StoreController extends AbstractController
     /**
     * @Route("/", name="home")
     */
-    public function index(ProductRepository $prodRepo): Response
+    public function home(ProductRepository $prodRepo): Response
     {
-        // $produits = $prodRepo->findAll();
-
-        return $this->render('index.html.twig', [
-            // 'produits' => $produits,
-        ]);
+    
+        return $this->render('index.html.twig');
     }
 
     /**
@@ -51,6 +51,10 @@ class StoreController extends AbstractController
     * @Route("/produit/edit/{id}", name="prod-edit")
     */
     public function addOrUpdateProduit(Product $product=null, Request $req, EntityManagerInterface $em){
+
+        if (!$this->isGranted('ROLE_ADMIN')) {
+			return $this->render('/erreur/accessDenied.html.twig');
+		}
         
         if(!$product){
             $product = new Product;
@@ -76,6 +80,10 @@ class StoreController extends AbstractController
     */
     public function deleteProduit(Product $product, EntityManagerInterface $em){
 
+        if (!$this->isGranted('ROLE_ADMIN')) {
+			return $this->render('/erreur/accessDenied.html.twig');
+		}
+
         if($product){
             $em->remove($product);
             $em->flush();
@@ -83,62 +91,4 @@ class StoreController extends AbstractController
 
         return $this->redirectToRoute("prods");
     }
-
-    // **********************************************
-    // *********************CATEGORIE****************   
-    // **********************************************
-
-
-    /**
-    * @Route("/category/view", name="cats")
-    */
-    public function catView(CategoryRepository $catRepo): Response
-    {
-        $categories = $catRepo->findAll();
-
-        return $this->render('store/cat.html.twig', [
-            'categories' => $categories,
-        ]);
-    }
-
-    /**
-    * @Route("/categorie/new", name="cat-new")
-    * @Route("/categorie/edit/{id}", name="cat-edit")
-    */
-    public function addOrUpdateCategorie(Category $category=null, Request $req, EntityManagerInterface $em){
-        
-        if(!$category){
-            $category = new category;
-        }
-
-        $formCategory = $this->createForm(CategoryType::class, $category);
-
-        $formCategory->handleRequest($req);
-
-        if ($formCategory->isSubmitted() && $formCategory->isValid()) {
-            $em->persist($category);
-            $em->flush();
-            return $this->redirectToRoute('prods',['id'=>$category->getId()]);
-        }
-        return $this->render('store/catForm.html.twig',[
-            'formCategory'=> $formCategory->createView(),
-            'mode'=> $category->getId() !== null
-        ]);
-    }
-
-    /**
-    * @Route("/categorie/delete/{id}", name="cat-delete")
-    */
-    public function deleteCategorie(Category $category, EntityManagerInterface $em){
-
-        if($category){
-            $em->remove($category);
-            $em->flush();
-        }
-
-        return $this->redirectToRoute("cats");
-    }
-
-
-
 }
